requirements.txt
Streamlit
pillow
google.generativeai